Add script for importing from csv file into database
The original data are stored in an excel file. There were problems with
utf8 using Spreadsheet::XLSX. The Excel file is in 2007 format. I changed
to importing the data from csv file format, which can be easily converted
with Excel. Because the data are stored in one flat table references
and links have to be extracted and stored in extra tables.

The Edit view is added with data of references and links. In the
controller data from has_many and many_to_many relationships are
flattened in an arrayref of hashrefs.